+++ src/Parser/Php/PhpAttributeTargetKind.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Symfony\Lsp\Parser\Php;
+
+enum PhpAttributeTargetKind
+{
+    case Type;
+    case Method;
+    case Property;
+}